Update User Type (JWT)
