+++ src/auth/auth.controller.ts
@@ -1,0 +1,33 @@
+import {
+  Controller,
+  Get,
+  Post,
+  Render,
+  Request,
+  UseGuards,
+} from '@nestjs/common';
+import { AuthService } from './auth.service';
+import { Public } from './decorator/customize';
+import { LocalAuthGuard } from './passport/local-auth.guard';
+
+@Controller()
+export class AuthController {
+  constructor(private authService: AuthService) {}
+  @Public()
+  @UseGuards(LocalAuthGuard)
+  @Post('auth/login')
+  async login(@Request() req) {
+    return this.authService.login(req.user);
+  }
+
+  @Get('profile')
+  async getProfile(@Request() req) {
+    return req.user;
+  }
+
+  @Get()
+  @Render('index')
+  getHello() {
+    // return this.appService.getHello();
+  }
+}
