+++ src/users/users.interface.ts
@@ -1,0 +1,6 @@
+export interface IUser {
+  _id: string;
+  name: string;
+  email: string;
+  role: string;
+}
